normalize path in view manager + remove warnings

+++ CodeMapDelta/CodeMapGUI/common_types.h
@@ -1,6 +1,6 @@
 #ifndef COMMON_TYPES_H
 #define COMMON_TYPES_H
 
-const size_t INVALID_INDEX = -1;
+const long long INVALID_INDEX = -1;
 
 #endif COMMON_TYPES_H
